refactor(drawer): renamed file to DrawerMenu
