+++ covid-mapper/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { store } from './store';
-import DrawerMenu from './features/drawer/Drawer'
+import DrawerMenu from './features/drawer/DrawerMenu'
 
 export default function App() {
   return (
